test: move proposal tests to its own file
